Update ViewPayments
This clever query gets invoice payments without searching for all the types of payments that can be made.

It uses the journal entry and internal reconciliation tables to get the data.

It displays the events of payment for all installments of the invoice. If the payment is not made for an Finance account (e.g credit card), it displays the value payed only when the deposit is made. If the deposit is cancelled, it displays zero and the date of the last event that was made. 

It displays the cancellations events for the payment too.

Some improments that can be made:
The applyed value formula can bem improved so it doesn't have to use the case statements, leaving all in just one formula (if possible).
Show the events for every transaction made (such as if a deposit is canceled than made again). Today only the applied value and the last date of the event is displayed. In the other way, like it is today, it is cleaner, because it doesn't show "unnecessary" rows.